Inserção da função status do equipamento em EquipmentMap.tsx
